Added a simple i2c dump.
The micro sends a pin high which interrupts the pi which then reads out
the i2c data. Once the data has been read the pi then writes 0xFF to
register 0xFF (which causes the micro to clear its buffer and set the
interrupt pin low) the program then waits for another interrupt to dump
the data again.

I still need to proccess this data but it shows that I can get the data
that is being send to the MRF radio with a little help from a atmega
micro.

Atm the dump process takes 370ms (The scope helps out again) and without
printing the contents to console the dump takes 356ms. This is due to the
slow baud rate I have the i2c bus configured too. When I was testing at
higher speeds with i2cdump I was getting intermittent errors so I lowered
the speed. I suspect this is due to the breadboard prototyping and once I
get this on some protoboard I will be able to increase the baud rate and
reduce the time it takes to read the data from the micro. The 350ms~ read
speed isn't bad or world ending I just feel I can inprove the speed.