Added a simple reset pulse to the micro on app start.
ATM if the micro gets data from the radio it will "lock" until the pi
queries it, which the pi might not do if the micro is already waiting for
the pi. So lets just disgrad any dat the micro has before we run the app.
I will recode the micro not to the buffer any data until the pi "Starts"
the micro, but for now these few lines of code will save any headaches.