Bugs bugs and more bugs...
Squeshing bugs left and right and up and down...
I don't like bugs, they are annoying.

Not actually fixed the bug yet, the bug seems to be in the micro code. So
I made a few changes that seem to of migrated the issues the bugs cause. I
need to keep an eye on it.